+++ src/solc/standard_json/output/source.rs
@@ -18,6 +18,7 @@ pub struct Source {
     /// The source code ID.
     pub id: usize,
     /// The source code AST.
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub ast: Option<serde_json::Value>,
 }