+++ src/tests/mod.rs
@@ -68,7 +68,7 @@ pub fn build_solidity(
     )?;
 
     let mut solc_output =
-        solc_compiler.standard_json(solc_input, Some(pipeline), None, vec![], None)?;
+        solc_compiler.standard_json(solc_input, Some(pipeline), vec![], None, vec![], None)?;
     solc_output.check_errors()?;
 
     let project = Project::try_from_solidity_sources(
@@ -139,7 +139,7 @@ pub fn build_solidity_and_detect_missing_libraries(
     )?;
 
     let mut solc_output =
-        solc_compiler.standard_json(solc_input, Some(pipeline), None, vec![], None)?;
+        solc_compiler.standard_json(solc_input, Some(pipeline), vec![], None, vec![], None)?;
 
     let project = Project::try_from_solidity_sources(
         sources,
@@ -174,7 +174,7 @@ pub fn build_yul(sources: BTreeMap<String, String>) -> anyhow::Result<SolcStanda
     let zksolc_version = semver::Version::parse(env!("CARGO_PKG_VERSION")).expect("Always valid");
     let optimizer_settings = era_compiler_llvm_context::OptimizerSettings::none();
 
-    let mut solc_output = SolcStandardJsonOutput::new(&sources);
+    let mut solc_output = SolcStandardJsonOutput::new(&sources, vec![]);
 
     let project = Project::try_from_yul_sources(
         sources,
@@ -222,10 +222,10 @@ pub fn build_yul_standard_json(
     let sources = solc_input.sources()?;
     let (solc_version, mut solc_output) = match solc_compiler {
         Some(solc_compiler) => {
-            let solc_output = solc_compiler.validate_yul_standard_json(solc_input)?;
+            let solc_output = solc_compiler.validate_yul_standard_json(solc_input, vec![])?;
             (Some(&solc_compiler.version), solc_output)
         }
-        None => (None, SolcStandardJsonOutput::new(&sources)),
+        None => (None, SolcStandardJsonOutput::new(&sources, vec![])),
     };
 
     let project = Project::try_from_yul_sources(
@@ -269,7 +269,7 @@ pub fn build_llvm_ir_standard_json(
     )?;
 
     let sources = solc_input.sources()?;
-    let mut solc_output = SolcStandardJsonOutput::new(&sources);
+    let mut solc_output = SolcStandardJsonOutput::new(&sources, vec![]);
 
     let project = Project::try_from_llvm_ir_sources(sources, Some(&mut solc_output))?;
     let build = project.compile_to_eravm(
@@ -306,7 +306,7 @@ pub fn build_eravm_assembly_standard_json(
     )?;
 
     let sources = solc_input.sources()?;
-    let mut solc_output = SolcStandardJsonOutput::new(&sources);
+    let mut solc_output = SolcStandardJsonOutput::new(&sources, vec![]);
 
     let project = Project::try_from_eravm_assembly_sources(sources, Some(&mut solc_output))?;
     let build = project.compile_to_eravm(
@@ -362,7 +362,7 @@ pub fn check_solidity_warning(
         suppressed_warnings,
     )?;
 
-    let solc_output = solc.standard_json(solc_input, Some(pipeline), None, vec![], None)?;
+    let solc_output = solc.standard_json(solc_input, Some(pipeline), vec![], None, vec![], None)?;
     let contains_warning = solc_output
         .errors
         .ok_or_else(|| anyhow::anyhow!("Solidity compiler messages not found"))?